pyproject.toml
[project]
name = "langchain-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain[openai]>=1.2.17",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langsmith>=0.8.0",
    "langchain-core>=1.3.2",
    "python-dotenv>=1.2.2",
    "jupyterlab>=4.5.7",
]
